+++ docs/docs.go
@@ -98,7 +98,7 @@ const docTemplate = `{
                         "in": "body",
                         "required": true,
                         "schema": {
-                            "$ref": "#/definitions/schemas.UnifiedChatRequest"
+                            "$ref": "#/definitions/schemas.ChatRequest"
                         }
                     }
                 ],
@@ -106,7 +106,7 @@ const docTemplate = `{
                     "200": {
                         "description": "OK",
                         "schema": {
-                            "$ref": "#/definitions/schemas.UnifiedChatResponse"
+                            "$ref": "#/definitions/schemas.ChatResponse"
                         }
                     },
                     "400": {
@@ -253,6 +253,52 @@ const docTemplate = `{
                 },
                 "user": {
                     "type": "string"
+                }
+            }
+        },
+        "bedrock.Config": {
+            "type": "object",
+            "required": [
+                "awsRegion",
+                "baseUrl",
+                "chatEndpoint",
+                "model"
+            ],
+            "properties": {
+                "awsRegion": {
+                    "type": "string"
+                },
+                "baseUrl": {
+                    "type": "string"
+                },
+                "chatEndpoint": {
+                    "type": "string"
+                },
+                "defaultParams": {
+                    "$ref": "#/definitions/bedrock.Params"
+                },
+                "model": {
+                    "type": "string"
+                }
+            }
+        },
+        "bedrock.Params": {
+            "type": "object",
+            "properties": {
+                "max_tokens": {
+                    "type": "integer"
+                },
+                "stop": {
+                    "type": "array",
+                    "items": {
+                        "type": "string"
+                    }
+                },
+                "temperature": {
+                    "type": "number"
+                },
+                "top_p": {
+                    "type": "number"
                 }
             }
         },
@@ -431,6 +477,84 @@ const docTemplate = `{
                 }
             }
         },
+        "ollama.Config": {
+            "type": "object",
+            "required": [
+                "baseUrl",
+                "chatEndpoint",
+                "model"
+            ],
+            "properties": {
+                "baseUrl": {
+                    "type": "string"
+                },
+                "chatEndpoint": {
+                    "type": "string"
+                },
+                "defaultParams": {
+                    "$ref": "#/definitions/ollama.Params"
+                },
+                "model": {
+                    "type": "string"
+                }
+            }
+        },
+        "ollama.Params": {
+            "type": "object",
+            "properties": {
+                "microstat": {
+                    "type": "integer"
+                },
+                "microstat_eta": {
+                    "type": "number"
+                },
+                "microstat_tau": {
+                    "type": "number"
+                },
+                "num_ctx": {
+                    "type": "integer"
+                },
+                "num_gpu": {
+                    "type": "integer"
+                },
+                "num_gqa": {
+                    "type": "integer"
+                },
+                "num_predict": {
+                    "type": "integer"
+                },
+                "num_thread": {
+                    "type": "integer"
+                },
+                "repeat_last_n": {
+                    "type": "integer"
+                },
+                "seed": {
+                    "type": "integer"
+                },
+                "stop": {
+                    "type": "array",
+                    "items": {
+                        "type": "string"
+                    }
+                },
+                "stream": {
+                    "type": "boolean"
+                },
+                "temperature": {
+                    "type": "number"
+                },
+                "tfs_z": {
+                    "type": "number"
+                },
+                "top_k": {
+                    "type": "integer"
+                },
+                "top_p": {
+                    "type": "number"
+                }
+            }
+        },
         "openai.Config": {
             "type": "object",
             "required": [
@@ -517,6 +641,9 @@ const docTemplate = `{
                 "azureopenai": {
                     "$ref": "#/definitions/azureopenai.Config"
                 },
+                "bedrock": {
+                    "$ref": "#/definitions/bedrock.Config"
+                },
                 "client": {
                     "$ref": "#/definitions/clients.ClientConfig"
                 },
@@ -539,6 +666,9 @@ const docTemplate = `{
                 },
                 "octoml": {
                     "$ref": "#/definitions/octoml.Config"
+                },
+                "ollama": {
+                    "$ref": "#/definitions/ollama.Config"
                 },
                 "openai": {
                     "description": "Add other providers like",
@@ -627,6 +757,52 @@ const docTemplate = `{
                 }
             }
         },
+        "schemas.ChatRequest": {
+            "type": "object",
+            "properties": {
+                "message": {
+                    "$ref": "#/definitions/schemas.ChatMessage"
+                },
+                "messageHistory": {
+                    "type": "array",
+                    "items": {
+                        "$ref": "#/definitions/schemas.ChatMessage"
+                    }
+                },
+                "override": {
+                    "$ref": "#/definitions/schemas.OverrideChatRequest"
+                }
+            }
+        },
+        "schemas.ChatResponse": {
+            "type": "object",
+            "properties": {
+                "cached": {
+                    "type": "boolean"
+                },
+                "created": {
+                    "type": "integer"
+                },
+                "id": {
+                    "type": "string"
+                },
+                "model": {
+                    "type": "string"
+                },
+                "modelResponse": {
+                    "$ref": "#/definitions/schemas.ProviderResponse"
+                },
+                "model_id": {
+                    "type": "string"
+                },
+                "provider": {
+                    "type": "string"
+                },
+                "router": {
+                    "type": "string"
+                }
+            }
+        },
         "schemas.OverrideChatRequest": {
             "type": "object",
             "properties": {
@@ -666,52 +842,6 @@ const docTemplate = `{
                 },
                 "totalTokens": {
                     "type": "number"
-                }
-            }
-        },
-        "schemas.UnifiedChatRequest": {
-            "type": "object",
-            "properties": {
-                "message": {
-                    "$ref": "#/definitions/schemas.ChatMessage"
-                },
-                "messageHistory": {
-                    "type": "array",
-                    "items": {
-                        "$ref": "#/definitions/schemas.ChatMessage"
-                    }
-                },
-                "override": {
-                    "$ref": "#/definitions/schemas.OverrideChatRequest"
-                }
-            }
-        },
-        "schemas.UnifiedChatResponse": {
-            "type": "object",
-            "properties": {
-                "cached": {
-                    "type": "boolean"
-                },
-                "created": {
-                    "type": "integer"
-                },
-                "id": {
-                    "type": "string"
-                },
-                "model": {
-                    "type": "string"
-                },
-                "modelResponse": {
-                    "$ref": "#/definitions/schemas.ProviderResponse"
-                },
-                "model_id": {
-                    "type": "string"
-                },
-                "provider": {
-                    "type": "string"
-                },
-                "router": {
-                    "type": "string"
                 }
             }
         }